events: Move sanitize HTML features to new ruma-html crate
