+++ crates/ruma-common/Cargo.toml
@@ -24,6 +24,7 @@ server = []
 api = ["dep:http", "dep:konst"]
 canonical-json = []
 events = []
+html = ["dep:ruma-html"]
 js = ["dep:js-sys", "getrandom?/js", "uuid?/js"]
 markdown = ["pulldown-cmark"]
 rand = ["dep:rand", "dep:uuid"]
@@ -47,7 +48,6 @@ unstable-msc3954 = ["unstable-msc1767"]
 unstable-msc3955 = ["unstable-msc1767"]
 unstable-msc3956 = ["unstable-msc1767"]
 unstable-pdu = []
-unstable-sanitize = ["dep:html5ever", "dep:phf"]
 unstable-unspecified = []
 
 # Don't validate the version part in `KeyId`.
@@ -77,17 +77,16 @@ base64 = { workspace = true }
 bytes = "1.0.1"
 form_urlencoded = "1.0.0"
 getrandom = { version = "0.2.6", optional = true }
-html5ever = { version = "0.26.0", optional = true }
 http = { workspace = true, optional = true }
 indexmap = { version = "2.0.0", features = ["serde"] }
 js_int = { workspace = true, features = ["serde"] }
 js_option = "0.1.0"
 konst = { version = "0.3.5", default-features = false, features = ["cmp", "iter", "parsing"], optional = true }
 percent-encoding = "2.1.0"
-phf = { version = "0.11.1", features = ["macros"], optional = true }
 pulldown-cmark = { version = "0.9.1", default-features = false, optional = true }
 rand = { version = "0.8.3", optional = true }
 regex = { version = "1.5.6", default-features = false, features = ["std", "perf"] }
+ruma-html = { workspace = true, optional = true }
 ruma-identifiers-validation = { workspace = true }
 ruma-macros = { workspace = true }
 serde = { workspace = true }
