+++ crates/ruma-html/CHANGELOG.md
@@ -1,0 +1,3 @@
+# 0.1.0 (unreleased)
+
+Initial release
