+++ crates/ruma-html/README.md
@@ -1,0 +1,12 @@
+# ruma-html
+
+[![crates.io page](https://img.shields.io/crates/v/ruma-html.svg)](https://crates.io/crates/ruma-html)
+[![docs.rs page](https://docs.rs/ruma-html/badge.svg)](https://docs.rs/ruma-html/)
+![license: MIT](https://img.shields.io/crates/l/ruma-html.svg)
+
+Opinionated HTML parsing and manipulating library.
+
+Like the rest of the Ruma crates, this crate is primarily meant to be used for
+the Matrix protocol. It should be able to be used to interact with any HTML
+content but will offer APIs focused on specificities of HTML in the Matrix
+specification.
